App/requirements.txt
streamlit
pycaret
pandas
plotly
numpy
streamlit-aggrid
pillow
